requirements.txt
beautifulsoup4==4.5.3
bleach==1.5.0
Django==1.10.5
djangorestframework==3.6.2
django-appconf==0.6
django-autoslug==1.9.3
django-debug-toolbar==1.6
django-html-sanitizer==0.1.5
django-jsonfield==0.9.13
django-mptt==0.8.7
django-pagedown==0.1.1
django-postman==3.1.0
django-registration==1.0
django-summernote==0.2.3
-e git://github.com/AASHE/django-tagulous@4f7dd6c04be0c70ccb5e23ca5652815bf3bb0258#egg=django_tagulous-develop
docutils==0.11
easy-thumbnails==1.3
html5lib==0.9999999
mpttadmin==0.3.1
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
python-dateutil==2.1
requests==2.13.0
six==1.10.0
sqlparse==0.2.2
wheel==0.24.0
django-allauth
steem
bcrypt
unidecode
googlemaps==2.4.6
django-extra-fields==0.9
transliterate==1.9
djangorestframework-jwt==1.10.0
django-filter==1.0.2
djangorestframework-recursive==0.1.1
draceditor==1.1.8
pyodbc==4.0.16
redis==2.10.5
redislite==3.2.311
psutil==5.2.2
lxml==3.8.0
coloredlogs==7.0
piston-cli==0.5.1
piston-lib==0.5.4
robot-detection==0.4
alphabet-detector==0.0.7
python-magic==0.4.13
bitshares==0.1.7
python-swiftclient==3.3.0
python-keystoneclient==3.12.0
python-novaclient==9.0.1
django-dbbackup==3.1.3
